radar_server

Function main

pub(crate) fn main()

Main handles argument parsing and calling the necessary submodules.

The main function takes the following arguments:
-h, –help Print this help message and exit
-s, –server Run in Server mode
-d, –dummy Run in Dummy Server mode
-a, –archive Run in Archiver mode
-t, –transform Run in Transformer mode
currently only Dummy mode and archiver mode do anything.

Server Mode

Sources real radar data and sends it to the DATA_ADDRESS.

ToDo: needs to be implemented

Dummy Mode

Sources simulated radar data and sends it to the DATA_ADDRESS.

Archive Mode

Sources radar packets from the SUBSCRIPTION_ADDRESS, archives each packet and forwards the live data the DATA_ADDRESS. Additionally will send archived data to individual clients on request.

ToDo: Sending archived data as well as forwarding live data is not yet implemented

ToDo: Additionally, should run on multiple threads, 1 to receive from a server instance, 1 to forward live data, 1 to handle communicating with a transformer instance, and 1 for each client wanting archived data.

Transform Mode

Sources data from an archiver instance of this program, transforms it based on the transformation requested by the client.

ToDo: needs to be implemented